requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
twilio
pytest
python-dotenv
pydantic
python-multipart
httpx